config: Extract `BindConfig` struct
This moves the `ip` and `port` settings from `Server` onto a dedicated
`BindConfig` struct in `config::bind`, along with the IP address selection
logic.
